chore: update dependence relation
