+++ packages/components/tooltip/__tests__/trigger.test.tsx
@@ -2,9 +2,9 @@ import { nextTick } from 'vue'
 import { shallowMount } from '@vue/test-utils'
 import { afterEach, describe, expect, it } from 'vitest'
 import { EVENT_CODE } from '@element-plus/constants'
+import { TOOLTIP_INJECTION_KEY } from '@element-plus/tokens'
 import ElTooltipTrigger from '../src/trigger.vue'
 import { genTooltipProvides } from '../test-helper/provides'
-import { TOOLTIP_INJECTION_KEY } from '../src/tokens'
 
 const AXIOM = 'rem is the best girl'
