--- packages/components/tooltip/src/tokens.ts
@@ -1,20 +1,0 @@
-import type { InjectionKey, Ref } from 'vue'
-import type { Trigger } from './tooltip'
-
-export type ElTooltipInjectionContext = {
-  controlled: Ref<boolean>
-  id: Ref<string>
-  open: Ref<boolean>
-  trigger: Ref<Trigger | Trigger[]>
-  onOpen: (e?: Event) => void
-  onClose: (e?: Event) => void
-  onToggle: (e: Event) => void
-  onShow: () => void
-  onHide: () => void
-  onBeforeShow: () => void
-  onBeforeHide: () => void
-  updatePopper: () => void
-}
-
-export const TOOLTIP_INJECTION_KEY: InjectionKey<ElTooltipInjectionContext> =
-  Symbol('elTooltip')
