+++ packages/components/tooltip/src/tooltip.ts
@@ -1,75 +1,16 @@
-// @ts-nocheck
-import { buildProps, definePropType } from '@element-plus/utils'
+import { buildProps } from '@element-plus/utils'
+import { createModelToggleComposable } from '@element-plus/hooks'
 import {
-  usePopperContentProps,
-  usePopperTriggerProps,
+  usePopperArrowProps,
+  usePopperProps,
 } from '@element-plus/components/popper'
-import {
-  POPPER_CONTAINER_SELECTOR,
-  useDelayedToggleProps,
-  useNamespace,
-} from '@element-plus/hooks'
-import { EVENT_CODE } from '@element-plus/constants'
-import { createModelToggleComposable } from '../../../hooks/use-model-toggle/index'
-import type Tooltip from '../tooltip.vue'
+import { useTooltipContentProps } from './content'
+import { useTooltipTriggerProps } from './trigger'
+import type { ElTooltipTriggerProps } from './trigger'
+import type { ElTooltipContentProps } from './content'
+import type Tooltip from './tooltip.vue'
 
 import type { ExtractPropTypes } from 'vue'
-
-const triggers = ['hover', 'focus', 'click', 'contextmenu'] as const
-const ns = useNamespace('tooltip')
-
-export type Trigger = typeof triggers[number]
-
-export const useTooltipContentProps = buildProps({
-  ...useDelayedToggleProps,
-  ...usePopperContentProps,
-  appendTo: {
-    type: definePropType<string | HTMLElement>([String, Object]),
-    default: POPPER_CONTAINER_SELECTOR,
-  },
-  content: {
-    type: String,
-    default: '',
-  },
-  rawContent: {
-    type: Boolean,
-    default: false,
-  },
-  persistent: Boolean,
-  ariaLabel: String,
-  // because model toggle prop is generated dynamically
-  // so the typing cannot be evaluated by typescript as type:
-  // [name]: { type: Boolean, default: null }
-  // so we need to declare that again for type checking.
-  visible: {
-    type: definePropType<boolean | null>(Boolean),
-    default: null,
-  },
-  transition: {
-    type: String,
-    default: `${ns.namespace.value}-fade-in-linear`,
-  },
-  teleported: {
-    type: Boolean,
-    default: true,
-  },
-  disabled: {
-    type: Boolean,
-  },
-} as const)
-
-export const useTooltipTriggerProps = buildProps({
-  ...usePopperTriggerProps,
-  disabled: Boolean,
-  trigger: {
-    type: definePropType<Trigger | Trigger[]>([String, Array]),
-    default: 'hover',
-  },
-  triggerKeys: {
-    type: definePropType<string[]>(Array),
-    default: () => [EVENT_CODE.enter, EVENT_CODE.space],
-  },
-} as const)
 
 export const useTooltipProps = buildProps({
   openDelay: {
@@ -93,13 +34,24 @@ export const createModelToggleComposableResult = createModelToggleComposable(
   'visible' as const
 )
 
-export type ElTooltipContentProps = ExtractPropTypes<
-  typeof useTooltipContentProps
->
+export const TooltipProps = {
+  ...usePopperProps,
+  ...createModelToggleComposableResult.useModelToggleProps,
+  ...useTooltipContentProps,
+  ...useTooltipTriggerProps,
+  ...usePopperArrowProps,
+  ...useTooltipProps,
+}
 
-export type ElTooltipTriggerProps = ExtractPropTypes<
-  typeof useTooltipTriggerProps
->
+export const TooltipEmits = [
+  ...createModelToggleComposableResult.useModelToggleEmits,
+  'before-show',
+  'before-hide',
+  'show',
+  'hide',
+  'open',
+  'close',
+]
 
 export type ElTooltipProps = ExtractPropTypes<typeof useTooltipProps> &
   ElTooltipContentProps &
