+++ packages/components/tooltip/src/trigger.vue
@@ -16,13 +16,13 @@
     <slot />
   </el-popper-trigger>
 </template>
-<script setup lang="ts">
+<script lang="ts" setup>
 import { inject, ref, toRef, unref } from 'vue'
 import { ElPopperTrigger } from '@element-plus/components/popper'
 import { composeEventHandlers } from '@element-plus/utils'
 import { useNamespace } from '@element-plus/hooks'
-import { TOOLTIP_INJECTION_KEY } from './tokens'
-import { useTooltipTriggerProps } from './tooltip'
+import { TOOLTIP_INJECTION_KEY } from '@element-plus/tokens'
+import { useTooltipTriggerProps } from './trigger'
 import { whenTrigger } from './utils'
 
 defineOptions({
