+++ packages/tokens/index.ts
@@ -15,5 +15,6 @@ export * from './tabs'
 export * from './upload'
 export * from './experimental-features'
 export * from './popper'
+export * from './tooltip'
 export * from './tooltip-v2'
 export * from './date-picker'
